+++ NEWS.md
@@ -1,3 +1,8 @@
+# AzureRMR 2.3.5.9000
+
+- Add ability to specify user-defined functions in `build_template_definition`. Also allow changing the schema, content version and api profile.
+- Change maintainer email address.
+
 # AzureRMR 2.3.5
 
 - Fix a bug in printing the error message when a template deployment fails.